Update scalloped stroke
Removed `withScallopedStroke` as this only added the original path to
the plot. It is unlikely users will want this for this style of line.
This also negates the need for the `_mpl._subclass_with_normal`
monkey patch.

Improved handling of control points at the end point of the path. The
unit tangent vector is calculated based on if the path is a closed
polygon or not. Tests were updated/added for these scenarios.